use f32 channels (error only on char grid render)

+++ singularity_ui/src/winit_backend/char_grid_shader.wgsl
@@ -56,7 +56,7 @@ fn vs_main(
 
 // I think each group changes together, and groups that change least should get lowest number
 @group(0) @binding(0)
-var sdf_atlas: texture_2d_array<u8>;
+var sdf_atlas: texture_2d_array<f32>;
 @group(0) @binding(1)
 var sdf_sampler: sampler;
 // Yeah, this is kinda jank, but I guess it makes it quirky *mews shyly*
